Switch to using an interface
This will allow our users not to have to depend on a concrete
implementation
