+++ src/PurpleBooth/HtmlStripper.php
@@ -7,30 +7,14 @@ namespace PurpleBooth;
  *
  * @package PurpleBooth
  */
-class HtmlStripper
+interface HtmlStripper
 {
     /**
-     * This is the extension itself, it's mostly a small wrapper around the actual parser.
+     * Parse html into text
      *
      * @param string $html
      *
      * @return string
      */
-    public function toText($html)
-    {
-        $parser = new Parser();
-
-        $xmlParser = xml_parser_create();
-        xml_set_element_handler($xmlParser, [$parser, 'startElement'], [$parser, 'endElement']);
-        xml_set_character_data_handler($xmlParser, [$parser, 'characterData']);
-
-        $wrappedHtml  = "<root>$html</root>";
-        $returnStatus = xml_parse($xmlParser, $wrappedHtml, true);
-
-        if (!$returnStatus) {
-            return $html;
-        }
-
-        return $parser->getText();
-    }
+    public function toText($html);
 }
